+++ frontend/src/pages/info.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Navigate } from 'react-router-dom';
+import { Button } from '@equinor/eds-core-react';
+  
+export default function Home() {
+    return (
+        <div>
+          <h2>Welcome to FLOW</h2>
+        
+        </div>
+      );
+}
